requirements.txt
numpy>=1.24.0
rasterio>=1.3.0
requests>=2.28.0
pyyaml>=6.0
pillow>=10.0.0
pyproj>=3.4.0
pystac-client>=0.7.0
planetary-computer>=1.0.0
scipy>=1.10.0
